requirements.txt
# Scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1

# BigQuery
google-cloud-bigquery==3.21.0
google-auth==2.29.0
pandas==2.2.2
pandas-gbq==0.23.0
pyarrow==16.0.0

# dbt
dbt-core==1.8.3
dbt-bigquery==1.8.2

# Airflow (installed separately via Docker, listed here for reference)
# apache-airflow==2.9.1

# Utils
python-dotenv==1.0.1
tenacity==8.3.0
tqdm==4.66.4
